Making tertiary/secondary/trunk road text show up sooner and bigger
I commented out the original rules (for reference), and made the
following changes to the <rules> in the driving.xml file:

* Line 601: tertiary highways -- now colored lime green (originally
yellow)
* Line 620 and 657: secondary/_link highways -- now colored magenta/pink
(originally orange)
* Line 684: trunk highways -- now colored purple (originally green)

I made the following revisions:

* zoom-min: from 14 to 10. I believe this will make the text show up
sooner. We may want to adjust it back to 12 or 13, but I went extreme to
make sure it works that way.
* font-size: from 10 to 14. I believe this will make the text larger for
all roads. Again, we may wan to adjust it back to 11 or 12, but I again
went extreme.
* stroke: I modified the color of each of these roadways to make sure I
was correcting the right type of road and not leaving any variations out
(for example: there's both a "secondary" and "secondary_link" yellow
roadway. I think we want to fix both of them). We'll want to change this
back, but I wanted to make sure I was effecting the correct road types
and getting all of them.